requirements.txt
# PINN EM Solver — dependencies
# Tested with CUDA 11.8, Python 3.10

# Core
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# NVIDIA Modulus (install separately from NVIDIA NGC or pip)
# nvidia-modulus>=23.09
# nvidia-modulus-sym>=23.09

# Config & experiment tracking
hydra-core>=1.3.0
omegaconf>=2.3.0
mlflow>=2.4.0

# Visualisation & demo
streamlit>=1.25.0
matplotlib>=3.7.0
plotly>=5.15.0

# Dev / testing
pytest>=7.3.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.0.280
